put all headers in header.h file

+++ backend/processor/multimedia_processor/descriptors.h
@@ -1,11 +1,7 @@
 #pragma once
 
-#include <pmp/SurfaceMesh.h>
-#include <pmp/algorithms/DifferentialGeometry.h>
-#include "eigen_vectors.h"
-
-
 namespace descriptors
 {
-    Eigen::VectorXf D1(pmp::SurfaceMesh &mesh, int bins);
+    // Maybe rename to DistBarycenterToRandomVertices, or add comment explaining what it calculates
+    Eigen::VectorXi D1(pmp::SurfaceMesh &mesh, int bins);
 }
